Get rid of the non-test compiler warnings seen on Darwin with gcc4
git-svn-id: https://www.boxbackup.org/svn/box/trunk@356 9c52e268-1103-0410-960d-c6a436f6b474

+++ test/backupdiff/testbackupdiff.cpp
@@ -83,7 +83,7 @@ void make_file_of_zeros(const char *filename, size_t size)
 	fclose(f);
 	free(b);
 
-	TEST_THAT(TestGetFileSize(filename) == size);
+	TEST_THAT((size_t)TestGetFileSize(filename) == size);
 }
